+++ src/lib/djerba/helpers/provenance_helper/test/helper_test.py
@@ -4,6 +4,7 @@
 
 import logging
 import os
+import time
 import unittest
 from configparser import ConfigParser
 import djerba.core.constants as core_constants
@@ -16,6 +17,8 @@ class TestProvenanceHelper(TestBase):
 
     CORE = 'core'
     HELPER_NAME = 'provenance_helper'
+    SUBSET_MD5 = '62c0b00c42a352d9ce3c49aedb55e8e2'
+    INFO_MD5 = '406227ea4d73df43d0e4adcd02fc14cb'
     
     def test(self):
         data_dir = os.path.join(os.environ.get('DJERBA_TEST_DATA'), 'helpers', 'provenance')
@@ -29,21 +32,27 @@ class TestProvenanceHelper(TestBase):
         config.set(self.HELPER_NAME, 'root_sample_name', 'PANX_1500')
         config.set(self.HELPER_NAME, 'provenance_input_path', provenance_input)
         config = helper_main.configure(config)
-        out_path = os.path.join(self.tmp_dir, helper_main.PROVENANCE_OUTPUT)
-        self.assertTrue(os.path.exists(out_path))
-        self.assertEqual(self.getMD5_of_gzip_path(out_path),
-                         '62c0b00c42a352d9ce3c49aedb55e8e2')
+        subset_path = os.path.join(self.tmp_dir, helper_main.PROVENANCE_OUTPUT)
+        self.assertTrue(os.path.exists(subset_path))
+        self.assertEqual(self.getMD5_of_gzip_path(subset_path), self.SUBSET_MD5)
         sample_info_path = os.path.join(self.tmp_dir, core_constants.DEFAULT_SAMPLE_INFO)
         self.assertTrue(os.path.exists(sample_info_path))
-        self.assertEqual(self.getMD5(sample_info_path), '406227ea4d73df43d0e4adcd02fc14cb')
-        # TODO suppress console output from helper_main_2
-        loader_2 = helper_loader(log_level=logging.DEBUG)
-        helper_main_2 = loader_2.load(self.HELPER_NAME, ws)
-        with self.assertLogs(level=logging.DEBUG) as log_context:
-            helper_main_2.extract(config)
-        msg = 'DEBUG:djerba.core.configure:extract: provenance_subset.tsv.gz already '+\
-            'in workspace, will not overwrite'
-        self.assertIn(msg, log_context.output)
+        self.assertEqual(self.getMD5(sample_info_path), self.INFO_MD5)
+        subset_mod = os.path.getmtime(subset_path)
+        info_mod = os.path.getmtime(sample_info_path)
+        time.sleep(0.01) # delay to enable file modification time check
+        helper_main.extract(config) # should do nothing
+        self.assertEqual(os.path.getmtime(subset_path), subset_mod)
+        self.assertEqual(os.path.getmtime(sample_info_path), info_mod)
+        ws.remove_file(subset_path)
+        ws.remove_file(sample_info_path)
+        helper_main.extract(config) # should regenerate the files
+        self.assertTrue(os.path.exists(subset_path))
+        self.assertEqual(self.getMD5_of_gzip_path(subset_path), self.SUBSET_MD5)
+        sample_info_path = os.path.join(self.tmp_dir, core_constants.DEFAULT_SAMPLE_INFO)
+        self.assertTrue(os.path.exists(sample_info_path))
+        self.assertEqual(self.getMD5(sample_info_path), self.INFO_MD5)
+
 
 if __name__ == '__main__':
     unittest.main()